requirements.txt
altgraph==0.17.3
asgiref==3.6.0
auto-py-to-exe==2.37.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.6.2
bottle==0.12.25
bottle-websocket==0.2.9
buildozer==1.5.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clr-loader==0.2.6
colorama==0.4.6
crayons==0.4.0
cryptography==41.0.7
customtkinter==5.2.2
cx-Freeze==6.15.13
cx-Logging==3.1.0
darkdetect==0.8.0
distlib==0.3.6
Django==4.2
dnspython==2.4.2
docutils==0.20.1
Eel==0.16.0
filelock==3.11.0
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
future==0.18.3
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
lief==0.14.0
MarkupSafe==2.1.2
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
paramiko==3.4.0
pefile==2023.2.7
pexpect==4.9.0
piexif==1.1.3
Pillow==9.5.0
platformdirs==3.2.0
proxy-tools==0.1.0
ptyprocess==0.7.0
pyasn1==0.5.1
pycparser==2.21
pyfiglet==1.0.2
pygame==2.5.2
Pygments==2.17.2
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PyNaCl==1.5.0
pyparsing==3.1.1
PyPDF2==3.0.1
pypiwin32==223
python-nmap==0.7.1
python-whois==0.8.0
pythonnet==3.0.3
pytube==15.0.0
pywebview==4.4.1
pywin32==306
pywin32-ctypes==0.2.2
requests==2.30.0
rsa==4.9
sh==2.0.6
soupsieve==2.4.1
speedtest-cli==2.1.3
SQLAlchemy==2.0.18
sqlparse==0.4.4
stegano==0.11.2
tkinterhtml==0.7
tkinterweb==3.23.5
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.21.0
Werkzeug==2.3.6
whichcraft==0.6.1
whois==0.99.3
WTForms==3.1.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
